refactor(client): split the main.rs into files by subcommand
