+++ client/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod cli;
+mod get_root;
+mod validate;
